Further refine zoom behavior
While we want to keep the node size constant, the scene should be
centered around the zoom event. This we can zoom into a specific
area. We also want to keep the pan functionality working. This
required the translate attribute to be set.

Finally, we removed the drag behavior as this didn't work with the
new zoom behavior and I don't think we need it anyway.